Create class Services and move static method getIdsOfModified inside
